requirements.txt
openai
google-auth
google-auth-oauthlib
google-api-python-client
moviepy
Pillow
python-dotenv
